Fix mixed cross-compile/native build
Relying on the real compiler tool specified by the main make_and_trace script
doesn't work with mixed cross-compile/native builds (e.g., with U-Boot). In this
case, the actual compiler tool is specified in the make recipe and should be
preserved as is.

This change implements another approach that relies on modifying the PATH
variable rather than specifying the exact tool path. The child script now simply
removes the PATH component added by the main script and runs the command as it's
specified in the make recipe.